Create auth module and move over some code
