+++ src/bin/webservice.rs
@@ -11,6 +11,7 @@ use rocket::request::Form;
 use rocket::response::status::Custom;
 use rocket_contrib::JSON;
 use schani_userinfo::*;
+use schani_userinfo::auth::verify_password;
 use schani_userinfo::models::Setting;
 
 #[derive(Debug, FromForm)]
